requirements.txt
asgiref==3.2.3
autopep8==1.4.4
awsebcli==3.16.0
blessed==1.16.1
boto3==1.10.41
botocore==1.13.41
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.3.9
cryptography==2.8
Django==3.0
django-cors-headers==3.2.0
django-fernet-fields==0.6
django-filter==2.2.0
django-storages==1.8
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
pathspec==0.5.9
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==3.13
requests==2.20.1
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
apscheduler==3.6.3
